requirements.txt
tenacity!=8.4.0
langchain==0.3.9
neo4j==5.27.0
streamlit==1.35.0
langchainhub==0.1.21
langchain-neo4j==0.1.1
langchain-ollama
python-dotenv
langchain-groq